% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPatients.R
\name{extractPatients}
\alias{extractPatients}
\title{Extract counts of distinct patients with conditions by year and month}
\usage{
extractPatients(connection, cdmSchema = "main")
}
\arguments{
\item{connection}{DatabaseConnector connection object.}

\item{cdmSchema}{Character. Schema that contains OMOP CDM tables.}
}
\value{
A data.frame with columns:
\itemize{
\item condition_concept_id (integer)
\item concept_name (character)
\item year (integer)
\item month (integer)
\item patient_count (integer)
}
}
\description{
Queries OMOP CDM condition_occurrence (joined to concept) to compute distinct
patient counts per condition, aggregated by year and month.
}
\examples{
\dontrun{
connection_details <- Eunomia::getEunomiaConnectionDetails()
conn <- connect(connection_details)
df <- extractPatients(connection = conn, cdmSchema = "main")
DatabaseConnector::disconnect(conn)
}

}
