% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchShinyApp.R
\name{launchShinyApp}
\alias{launchShinyApp}
\title{Launch the omoptrend Shiny app with pre-extracted data}
\usage{
launchShinyApp(data)
}
\arguments{
\item{data}{data.frame produced by extractPatients()}
}
\description{
Launch the omoptrend Shiny app with pre-extracted data
}
