% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_validation.R
\name{validateConnection}
\alias{validateConnection}
\title{Validate a DatabaseConnector connection}
\usage{
validateConnection(connection)
}
\description{
Validate a DatabaseConnector connection
}
\keyword{internal}
