% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_validation.R
\name{validateDialect}
\alias{validateDialect}
\title{Validate supported dialect via CommonDataModel}
\usage{
validateDialect(targetDialect)
}
\description{
Validate supported dialect via CommonDataModel
}
\keyword{internal}
