% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_validation.R
\name{validateExtractedData}
\alias{validateExtractedData}
\title{Validate structure of extracted data}
\usage{
validateExtractedData(data)
}
\description{
Validate structure of extracted data
}
\keyword{internal}
