% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTrend.R
\name{plotTrend}
\alias{plotTrend}
\title{Plot condition patient counts by year or aggregated month}
\usage{
plotTrend(data, byMonth = FALSE)
}
\arguments{
\item{data}{data.frame from extractPatients()
Must include columns: condition_concept_id, concept_name, year, month, patient_count}

\item{byMonth}{Logical. TRUE = plot by month (aggregated across years),
FALSE = plot by year (aggregated across months).}
}
\value{
ggplot object showing patient count trends
}
\description{
The funciton takes output from the \code{extractPatients} function as input and returns a plot with the number of patients
per year (default) or per month for each condition in "data".
}
\examples{
\dontrun{
p <- plotTrend(data = df, byMonth = FALSE)
print(p)
}

\dontrun{
connection_details <- Eunomia::getEunomiaConnectionDetails()
conn <- connect(connection_details)

df <- extractPatients(connection = conn, cdmSchema = "main")

# Plot trends by YEAR
plotTrend(data = df)

# Plot trends by MONTH
plotTrend(data = df, byMonth = TRUE)
}

}
